+++ libraries/exoplayer/src/main/java/androidx/media3/exoplayer/source/preload/BasePreloadManager.java
@@ -18,10 +18,13 @@ package androidx.media3.exoplayer.source.preload;
 import static androidx.media3.common.util.Assertions.checkNotNull;
 
 import android.os.Handler;
+import android.os.Looper;
 import androidx.annotation.GuardedBy;
 import androidx.annotation.Nullable;
 import androidx.media3.common.C;
 import androidx.media3.common.MediaItem;
+import androidx.media3.common.util.Clock;
+import androidx.media3.common.util.ListenerSet;
 import androidx.media3.common.util.UnstableApi;
 import androidx.media3.common.util.Util;
 import androidx.media3.exoplayer.source.MediaSource;
@@ -58,10 +61,22 @@ public abstract class BasePreloadManager<T> {
     public abstract BasePreloadManager<T> build();
   }
 
+  /** Listener for events in a preload manager. */
+  public interface Listener {
+
+    /** Called when the given {@link MediaItem} has completed preloading. */
+    void onCompleted(MediaItem mediaItem);
+
+    /** Called when an {@linkplain PreloadException error} occurs. */
+    void onError(PreloadException exception);
+  }
+
   private final Object lock;
+  private final Looper looper;
   protected final Comparator<T> rankingDataComparator;
   private final TargetPreloadStatusControl<T> targetPreloadStatusControl;
   private final MediaSource.Factory mediaSourceFactory;
+  private final ListenerSet<Listener> listeners;
   private final Map<MediaItem, MediaSourceHolder> mediaItemMediaSourceHolderMap;
   private final Handler startPreloadingHandler;
 
@@ -77,12 +92,43 @@ public abstract class BasePreloadManager<T> {
       TargetPreloadStatusControl<T> targetPreloadStatusControl,
       MediaSource.Factory mediaSourceFactory) {
     lock = new Object();
+    looper = Util.getCurrentOrMainLooper();
     this.rankingDataComparator = rankingDataComparator;
     this.targetPreloadStatusControl = targetPreloadStatusControl;
     this.mediaSourceFactory = mediaSourceFactory;
+    listeners = new ListenerSet<>(looper, Clock.DEFAULT, (listener, flags) -> {});
     mediaItemMediaSourceHolderMap = new HashMap<>();
     startPreloadingHandler = Util.createHandlerForCurrentOrMainLooper();
     sourceHolderPriorityQueue = new PriorityQueue<>();
+  }
+
+  /**
+   * Adds a {@link Listener} to listen to the preload events.
+   *
+   * <p>This method can be called from any thread.
+   */
+  public void addListener(Listener listener) {
+    listeners.add(listener);
+  }
+
+  /**
+   * Removes a {@link Listener}.
+   *
+   * @throws IllegalStateException If this method is called from the wrong thread.
+   */
+  public void removeListener(Listener listener) {
+    verifyApplicationThread();
+    listeners.remove(listener);
+  }
+
+  /**
+   * Clears all the {@linkplain Listener listeners}.
+   *
+   * @throws IllegalStateException If this method is called from the wrong thread.
+   */
+  public void clearListeners() {
+    verifyApplicationThread();
+    listeners.clear();
   }
 
   /**
@@ -206,15 +252,33 @@ public abstract class BasePreloadManager<T> {
   public final void release() {
     reset();
     releaseInternal();
+    clearListeners();
   }
 
-  /** Called when the given {@link MediaSource} completes to preload. */
+  /** Called when the given {@link MediaSource} completes preloading. */
   protected final void onPreloadCompleted(MediaSource source) {
+    listeners.sendEvent(
+        /* eventFlag= */ C.INDEX_UNSET, listener -> listener.onCompleted(source.getMediaItem()));
+    maybeAdvanceToNextSource(source);
+  }
+
+  /** Called when an error occurs. */
+  protected final void onPreloadError(PreloadException error, MediaSource source) {
+    listeners.sendEvent(/* eventFlag= */ C.INDEX_UNSET, listener -> listener.onError(error));
+    maybeAdvanceToNextSource(source);
+  }
+
+  /** Called when the given {@link MediaSource} has been skipped before completing preloading. */
+  protected final void onPreloadSkipped(MediaSource source) {
+    maybeAdvanceToNextSource(source);
+  }
+
+  private void maybeAdvanceToNextSource(MediaSource preloadingSource) {
     startPreloadingHandler.post(
         () -> {
           synchronized (lock) {
             if (sourceHolderPriorityQueue.isEmpty()
-                || checkNotNull(sourceHolderPriorityQueue.peek()).mediaSource != source) {
+                || checkNotNull(sourceHolderPriorityQueue.peek()).mediaSource != preloadingSource) {
               return;
             }
             do {
@@ -305,6 +369,12 @@ public abstract class BasePreloadManager<T> {
       }
     }
     return false;
+  }
+
+  private void verifyApplicationThread() {
+    if (Looper.myLooper() != looper) {
+      throw new IllegalStateException("Preload manager is accessed on the wrong thread.");
+    }
   }
 
   /** A holder for information for preloading a single media source. */